requirements.txt
python-telegram-bot
requests
PyYAML
psycopg2-binary